sysvinit: upgrade 3.04 -> 3.10
Upgrade	of sysvinit to version 3.10. The upstream moved	in between to github.
Removed	the following patches:
- 0001-hddown-include-libgen.h-for-basename-API.patch
  Is already included in upstream
- install.patch
  Referenced PR is already merged in upstream
- sysvinit_remove_linux_fs.patch
  Is already included in upstream

The remaining patches were mostly just updated, with exception to the realpath.patch
Here the changes to the killall5.c source file couldn't be merged anymore.
My guess is, that the changes are already done someway in the upstream.
At least the file is now using realpath(). I didn't touch the remaining additional includes,
as I don't know if they are still needed.

Signed-off-by: Sven <mr.name@web.de>